VistorIA/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
openai==1.101.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
reportlab==4.4.3
sniffio==1.3.1
starlette==0.47.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
# Database
sqlalchemy==2.0.25
# sqlite3 já está incluído no Python

# Background processing
celery==5.3.6
redis==5.0.1

# OCR and Computer Vision
pytesseract==0.3.10
opencv-python==4.9.0.80
ultralytics==8.0.238

# Speech Recognition
speechrecognition==3.10.1
pyaudio==0.2.14

# Additional utilities
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
